Create a clearbutton.js file for ES6 modules
