+++ src/clearButton.js
@@ -1,0 +1,2 @@
+const clearButton = document.getElementById('clear-btn');
+export default clearButton;
